backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
alembic
pydantic-settings
opencv-python-headless
ultralytics 
labelImg
requests